updates for the pruner
git-svn-id: https://svn.jboss.org/repos/labs/labs/jbossrules/contrib@21142 c60d74c8-e8f6-0310-9e8f-d4a2fc68ab70

+++ machinelearning/5.0/drools-core/src/main/java/org/drools/learner/DecisionTree.java
@@ -30,7 +30,7 @@ public class DecisionTree {
 	private String execution_signature;
 	public long FACTS_READ = 0;
 
-	private int validation_error, training_error;
+	private double validation_error, training_error;
 
 	private int num_nonterminal_nodes;
 
@@ -116,17 +116,17 @@ public class DecisionTree {
 	}
 	
 
-	public void setValidationError(int error) {
+	public void setValidationError(double error) {
 		validation_error = error;
 	}	
-	public int getValidationError() {
+	public double getValidationError() {
 		return validation_error;
 	}
 	
-	public void setTrainingError(int error) {
+	public void setTrainingError(double error) {
 		training_error = error;
 	}
-	public int getTrainingError() {
+	public double getTrainingError() {
 		return training_error;
 	}
